requirements.txt
Flask
deepgram-sdk
yt-dlp
python-dotenv
gunicorn
